refactoring, merged Offsite Alipay and Internetbanking Payment Builder classes to one class
